+++ src/services/globalization/globalization.service.ts
@@ -8,10 +8,10 @@ export abstract class GlobalizationService
     /**
      * Gets current name of locale, that is used within application
      */
-    public abstract getLocale(): string;
+    public abstract get locale(): string;
     
     /**
-     * Gets observable that emits data when locale changes and change should be applied to application
+     * Gets observable that emits when locale changes and change should be applied to application
      */
-    public abstract getLocaleChange(): Observable<string>;
+    public abstract get localeChange(): Observable<void>;
 }